fixed the problems in master

+++ app/models/ticket.php
@@ -6,7 +6,7 @@ class Ticket implements \JsonSerializable
     private int $id;
     private string $title;
     private string $type;
-    private int $price;
+    private float $price;
     private Event $event;
 
     #[ReturnTypeWillChange]
@@ -47,7 +47,7 @@ class Ticket implements \JsonSerializable
     {
         return $this->price;
     }
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
         return $this;
